Switch to PSR-4, fix CS and update service provider
