+++ src/Interfaces/SearchDriverInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace TomLingham\Searchy\Interfaces;
+
+interface SearchDriverInterface
+{
+    public function query($searchString);
+
+    public function select(/* $columns */);
+
+    public function get();
+}
