+++ src/SearchDrivers/LevenshteinSearchDriver.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TomLingham\Searchy\SearchDrivers;
+
+class LevenshteinSearchDriver extends BaseSearchDriver
+{
+    /**
+     * @var array
+     */
+    protected $matchers = [
+        'TomLingham\Searchy\Matchers\LevenshteinMatcher' => 100,
+    ];
+}
