--- src/TomLingham/Searchy/Interfaces/SearchDriverInterface.php
@@ -1,12 +1,0 @@
-<?php namespace TomLingham\Searchy\Interfaces;
-
-interface SearchDriverInterface
-{
-
-	public function query( $searchString );
-
-	public function select( /* $columns */ );
-
-	public function get();
-
-}
