--- src/TomLingham/Searchy/Matchers/StartOfStringMatcher.php
@@ -1,28 +1,0 @@
-<?php namespace TomLingham\Searchy\Matchers;
-
-/**
- * Matches Strings that begin with the search string.
- *
- * For example, a search for 'hel' would match; 'Hello World' or 'helping hand'
- *
- * Class StartOfStringMatcher
- * @package TomLingham\Searchy\Matchers
- */
-
-class StartOfStringMatcher extends BaseMatcher
-{
-
-	/**
-	 * @var string
-	 */
-	protected $operator = 'LIKE';
-
-	/**
-	 * @param $searchString
-	 * @return string
-	 */
-	public function formatSearchString( $searchString )
-	{
-		return "$searchString%";
-	}
-}
