--- src/TomLingham/Searchy/SearchDrivers/LevenshteinSearchDriver.php
@@ -1,12 +1,0 @@
-<?php namespace TomLingham\Searchy\SearchDrivers;
-
-class LevenshteinSearchDriver extends BaseSearchDriver {
-
-	/**
-	 * @var array
-	 */
-	protected $matchers = [
-		'TomLingham\Searchy\Matchers\LevenshteinMatcher' => 100
-	];
-
-}
